requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0
openai==1.12.0
python-dotenv==1.0.0
sqlalchemy==2.0.23
schedule==1.2.0
python-dateutil==2.8.2
pytz==2023.3
pydantic==2.5.0
slowapi==0.1.9
python-multipart==0.0.6 